refactored and commented some components
