+++ Website/src/app/components/base_layout.tsx
@@ -1,0 +1,32 @@
+import './globals.css'
+import Header from "@/app/components/header";
+import Footer from "@/app/components/footer";
+
+/**
+ * The general layout for this site to be used both for pages
+ * using the app-router and the legacy pages-router.
+ *
+ * This will ensure that each page has a header and a footer.
+ *
+ * Because this is a flex-layout, children can use the `grow`
+ * class to grow to take up the remaining space on the page.
+ *
+ * @param children The actual page this layout should wrap. In JSX, these are the
+ *                 children of this element.
+ * @constructor
+ */
+export default function BaseLayout({
+    children,
+  }: {
+    children: React.ReactNode
+  }) {
+    return (
+        // inline styling using tailwind will keep styling information in the same
+        // file as the markup information.
+        <div className='h-full min-h-screen flex flex-initial flex-col'>
+            <Header/>
+            {children}
+            <Footer/>
+        </div>
+    )
+  }
