+++ Website/src/app/layout.tsx
@@ -1,15 +1,19 @@
-import Layout from "@/app/components/layout_base"
-import {inter, meta_info} from "@/lib/common";
+import BaseLayout from "@/app/components/base_layout"
+import {inter, meta_info} from "@/utils/common";
 
 export const metadata = meta_info;
 
+/**
+ * The Layout to use on all pages in the app-directory.
+ * Effectively defers to BaseLayout with minimal adjustments.
+ */
 export default function RootLayout({children,}: { children: React.ReactNode }) {
     return (
         <html lang="en">
             <body className={inter.className}>
-                <Layout>
+                <BaseLayout>
                     {children}
-                </Layout>
+                </BaseLayout>
             </body>
         </html>
     )
