+++ Website/src/app/login/page.tsx
@@ -1,15 +1,12 @@
-import Login from "@/components/login";
+import Login from "@/app/components/login";
+import {FormWrapper} from "@/app/components/form";
 
 
-export default function Home() {
-  
+export default function LoginPage() {
+
     return (
-            // <div className='h-full min-h-screen'>
-                <main className="container mx-auto max-w-2xl grow">
-                    <div className={'bg-white dark:bg-slate-800 dark:text-white p-4 rounded'}>
-                        <Login/>
-                    </div>
-                </main>
-            //</div>
+        <FormWrapper>
+            <Login/>
+        </FormWrapper>
     )
-  }
+}
