+++ Website/src/app/select_track/page.tsx
@@ -1,12 +1,10 @@
-import Selection from "@/components/track_selection";
+import Selection from "@/app/components/track_selection";
+import {FormWrapper} from "@/app/components/form";
 
 export default function Page() {
     return (
-        // <div className='h-full min-h-screen'>
-        <main className="container mx-auto max-w-2xl grow">
-            <div className={'bg-white dark:bg-slate-800 dark:text-white p-4 rounded'}>
-                <Selection/>
-            </div>
-        </main>
+        <FormWrapper>
+            <Selection/>
+        </FormWrapper>
     );
 }
